Move ChatPanelWindow to its own file
