+++ Core/Sources/SuggestionWidget/ChatPanelWindow.swift
@@ -1,0 +1,33 @@
+import AppKit
+import Foundation
+import SwiftUI
+
+final class ChatPanelWindow: NSWindow {
+    override var canBecomeKey: Bool { true }
+    override var canBecomeMain: Bool { true }
+
+    var minimizeWindow: () -> Void = {}
+
+    var isWindowHidden: Bool = false {
+        didSet {
+            alphaValue = isPanelDisplayed && !isWindowHidden ? 1 : 0
+        }
+    }
+
+    var isPanelDisplayed: Bool = false {
+        didSet {
+            alphaValue = isPanelDisplayed && !isWindowHidden ? 1 : 0
+        }
+    }
+
+    override var alphaValue: CGFloat {
+        didSet {
+            ignoresMouseEvents = alphaValue <= 0
+        }
+    }
+
+    override func miniaturize(_: Any?) {
+        minimizeWindow()
+    }
+}
+
